anazon/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-phonenumber-field==7.1.0
idna==3.4
phonenumbers==8.13.13
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.4.4
stripe==5.4.0
tzdata==2023.3
urllib3==2.0.3
